Remove custom external link counter CSS
This counter is provided by mediawiki.skinning.content.parsoid and we
should get it from there.  However, in T163738 it was removed from make-
css-assets.bash and the custom snippet here since other code in that
module was adversely affecting the Android app's image widening.

This removes the custom snippet again, and a separate Android app patch
will add mediawiki.skinning.content.parsoid to the app bundle and modify
the request so that image widening still works as intended.

Bug: T188919
Change-Id: I2ccd81a683002270cd5ec74431bf1d033d4373a0